Moved refrerencePointChache  to sentlog
