+++ pkg/backend/grpcServer/handlers/indirectHandler.go
@@ -9,13 +9,13 @@ import (
 
 type IndirectHandler struct {
 	sentLog *sentLog.SentLogServer
-	rpCache *ReferencePointCache
+	rpCache *sentLog.ReferencePointCache
 }
 
 func InitIndirectHandler(sLog *sentLog.SentLogServer) *IndirectHandler {
 	indirectHandler := &IndirectHandler{}
 	indirectHandler.sentLog = sLog
-	rpCache := InitReferencePointCache()
+	rpCache := sentLog.InitReferencePointCache()
 	rpCache.PopulateWithMockedData()
 	indirectHandler.rpCache = rpCache
