+++ pkg/backend/sentLog/referencePointCache_test.go
@@ -1,4 +1,4 @@
-package handlers
+package sentLog
 
 import (
 	"sync"
